refactor: file restructuring and implemented the forms and other asssets
