+++ public/assets/images/laptop.png
