+++ src/components/icons/HeartStrokeIcon.tsx
