+++ src/components/icons/StarIcon.tsx
