+++ src/components/product-card/index.tsx
@@ -1,5 +1,5 @@
-import CircularArrowIcon from "../../assets/icons/CircularArrowIcon";
-import HeartIcon from "../../assets/icons/HeartIcon";
+import CircularArrowIcon from "../icons/CircularArrowIcon";
+import HeartIcon from "../icons/HeartIcon";
 type TProps = {
   className?: string;
   name: string;
@@ -30,7 +30,9 @@ export default function ProductCard({
         </div>
         <p>{name}</p>
         <div className="w-full flex items-center space-x-10">
-          <span className="font-bold ">Rwf {price.toLocaleString()}</span>{" "}
+          <span className="font-bold okrika-bold ">
+            Rwf {price.toLocaleString()}
+          </span>{" "}
           <CircularArrowIcon />{" "}
         </div>
       </div>
