+++ src/layout/HomeLayout.tsx
@@ -1,9 +1,10 @@
-import HeartStrokeIcon from "../assets/icons/HeartStrokeIcon";
-import InstagramIcon from "../assets/icons/InstagramIcon";
-import ProfileIcon from "../assets/icons/ProfileIcon";
-import { SearhIcon } from "../assets/icons/SearchIcon";
-import WhatsappIcon from "../assets/icons/WhatsappIcon";
-import XIcon from "../assets/icons/XIcon";
+import HeartStrokeIcon from "../components/icons/HeartStrokeIcon";
+import InstagramIcon from "../components/icons/InstagramIcon";
+import ProfileIcon from "../components/icons/ProfileIcon";
+import { SearhIcon } from "../components/icons/SearchIcon";
+import WhatsappIcon from "../components/icons/WhatsappIcon";
+import XIcon from "../components/icons/XIcon";
+import Logo from "/assets/logo.svg";
 
 export default function HomeLayout({
   children
@@ -12,7 +13,7 @@ export default function HomeLayout({
     <div className="w-full pt-5 px-8">
       <header className="flex items-center justify-center">
         <nav className="w-[10%] flex rounded-[40px] shadow-lg md:w-4/5 lg:w-4/5 px-8 pb-7 pt-5 items-center justify-between">
-          <h1>Okrika</h1>
+          <img src={Logo} alt="Okrika" />
           <ul className="w-1/4 flex items-center text-lg space-x-8">
             <li>Home</li>
             <li>Categories</li>
@@ -40,7 +41,7 @@ export default function HomeLayout({
       </header>
       <main className="py-6">{children}</main>
       <footer className="flex flex-col bg-okrika-green items-center p-5 pt-10 rounded-lg  ">
-        <div className=" flex items-start w-[70%]">
+        <div className=" flex items-start w-[75%]">
           <div className=" flex items-center text-white flex-wrap">
             <span className="w-1/3">About Us</span>
             <span className="w-1/3">Terms and Conditions</span>
@@ -51,7 +52,10 @@ export default function HomeLayout({
             <span className="w-1/3">Community</span>
             <span className="w-1/3">Resources</span>
           </div>
-          <button className="bg-white text-black text-sm p-2 rounded-lg font-bold">
+          <button
+            type="button"
+            className="bg-white text-black text-sm p-2 rounded-lg okrika-bold"
+          >
             {" "}
             Shop Now
           </button>
@@ -59,7 +63,7 @@ export default function HomeLayout({
         <div className="w-full flex items-end justify-between">
           {" "}
           <div className="w-1/8 flex flex-col items-center space-y-4">
-            <h1>Okrika</h1>
+            <img src={Logo} alt="Okrika" />
             <div className="flex items-center space-x-4">
               <span>
                 <WhatsappIcon />
